+++ tests/Equinox.CosmosStore.Integration/CosmosIntegration.fs
@@ -1,8 +1,8 @@
-﻿module Equinox.Cosmos.Integration.CosmosIntegration
+﻿module Equinox.CosmosStore.Integration.CosmosIntegration
 
 open Domain
-open Equinox.Cosmos
-open Equinox.Cosmos.Integration.Infrastructure
+open Equinox.CosmosStore
+open Equinox.CosmosStore.Integration.Infrastructure
 open FSharp.UMX
 open Swensen.Unquote
 open System
@@ -12,45 +12,40 @@ module Cart =
     let fold, initial = Domain.Cart.Fold.fold, Domain.Cart.Fold.initial
     let snapshot = Domain.Cart.Fold.isOrigin, Domain.Cart.Fold.snapshot
     let codec = Domain.Cart.Events.codec
-    let createServiceWithoutOptimization connection batchSize log =
-        let store = createCosmosContext connection batchSize
-        let resolve (id,opt) = Resolver(store, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.Unoptimized).Resolve(id,?option=opt)
+    let createServiceWithoutOptimization log store =
+        let resolve (id,opt) = CosmosStoreCategory(store, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.Unoptimized).Resolve(id,?option=opt)
         Backend.Cart.create log resolve
     let projection = "Compacted",snd snapshot
     /// Trigger looking in Tip (we want those calls to occur, but without leaning on snapshots, which would reduce the paths covered)
-    let createServiceWithEmptyUnfolds connection batchSize log =
-        let store = createCosmosContext connection batchSize
+    let createServiceWithEmptyUnfolds store log =
         let unfArgs = Domain.Cart.Fold.isOrigin, fun _ -> Seq.empty
-        let resolve (id,opt) = Resolver(store, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.MultiSnapshot unfArgs).Resolve(id,?option=opt)
+        let resolve (id,opt) = CosmosStoreCategory(store, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.MultiSnapshot unfArgs).Resolve(id,?option=opt)
         Backend.Cart.create log resolve
-    let createServiceWithSnapshotStrategy connection batchSize log =
-        let store = createCosmosContext connection batchSize
-        let resolve (id,opt) = Resolver(store, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.Snapshot snapshot).Resolve(id,?option=opt)
+    let createServiceWithSnapshotStrategy store log =
+        let resolve (id,opt) = CosmosStoreCategory(store, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.Snapshot snapshot).Resolve(id,?option=opt)
         Backend.Cart.create log resolve
-    let createServiceWithSnapshotStrategyAndCaching connection batchSize log cache =
-        let store = createCosmosContext connection batchSize
+    let createServiceWithSnapshotStrategyAndCaching store log cache =
         let sliding20m = CachingStrategy.SlidingWindow (cache, TimeSpan.FromMinutes 20.)
-        let resolve (id,opt) = Resolver(store, codec, fold, initial, sliding20m, AccessStrategy.Snapshot snapshot).Resolve(id,?option=opt)
+        let resolve (id,opt) = CosmosStoreCategory(store, codec, fold, initial, sliding20m, AccessStrategy.Snapshot snapshot).Resolve(id,?option=opt)
         Backend.Cart.create log resolve
-    let createServiceWithRollingState connection log =
-        let store = createCosmosContext connection 1
+    let createServiceWithRollingState store log =
         let access = AccessStrategy.RollingState Domain.Cart.Fold.snapshot
-        let resolve (id,opt) = Resolver(store, codec, fold, initial, CachingStrategy.NoCaching, access).Resolve(id,?option=opt)
+        let resolve (id,opt) = CosmosStoreCategory(store, codec, fold, initial, CachingStrategy.NoCaching, access).Resolve(id,?option=opt)
         Backend.Cart.create log resolve
 
 module ContactPreferences =
     let fold, initial = Domain.ContactPreferences.Fold.fold, Domain.ContactPreferences.Fold.initial
     let codec = Domain.ContactPreferences.Events.codec
-    let createServiceWithoutOptimization createGateway defaultBatchSize log _ignoreWindowSize _ignoreCompactionPredicate =
-        let gateway = createGateway defaultBatchSize
-        let resolver = Resolver(gateway, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.Unoptimized)
-        Backend.ContactPreferences.create log resolver.Resolve
-    let createService log createGateway =
-        let resolver = Resolver(createGateway 1, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.LatestKnownEvent)
-        Backend.ContactPreferences.create log resolver.Resolve
-    let createServiceWithLatestKnownEvent createGateway log cachingStrategy =
-        let resolver = Resolver(createGateway 1, codec, fold, initial, cachingStrategy, AccessStrategy.LatestKnownEvent)
-        Backend.ContactPreferences.create log resolver.Resolve
+    let createServiceWithoutOptimization createContext defaultBatchSize log _ignoreWindowSize _ignoreCompactionPredicate =
+        let context = createContext defaultBatchSize
+        let resolve = CosmosStoreCategory(context, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.Unoptimized).Resolve
+        Backend.ContactPreferences.create log resolve
+    let createService log store =
+        let resolve = CosmosStoreCategory(store, codec, fold, initial, CachingStrategy.NoCaching, AccessStrategy.LatestKnownEvent).Resolve
+        Backend.ContactPreferences.create log resolve
+    let createServiceWithLatestKnownEvent store log cachingStrategy =
+        let resolve = CosmosStoreCategory(store, codec, fold, initial, cachingStrategy, AccessStrategy.LatestKnownEvent).Resolve
+        Backend.ContactPreferences.create log resolve
 
 #nowarn "1182" // From hereon in, we may have some 'unused' privates (the tests)
 
@@ -75,12 +70,12 @@ type Tests(testOutputHelper) =
         let tripRequestCharges = [ for e, c in capture.RequestCharges -> sprintf "%A" e, c ]
         test <@ float rus >= Seq.sum (Seq.map snd tripRequestCharges) @>
 
-    [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
+    [<AutoData(MaxFail=1, MaxTest=2, SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
     let ``Can roundtrip against Cosmos, correctly batching the reads [without reading the Tip]`` context skuId = Async.RunSynchronously <| async {
-        let! conn = connectToSpecifiedCosmosOrSimulator log
-
         let maxItemsPerRequest = 5
-        let service = Cart.createServiceWithoutOptimization conn maxItemsPerRequest log
+        let store = createPrimaryContext log maxItemsPerRequest
+
+        let service = Cart.createServiceWithoutOptimization log store
         capture.Clear() // for re-runs of the test
 
         let cartId = % Guid.NewGuid()
@@ -103,22 +98,22 @@ type Tests(testOutputHelper) =
 
         let expectedResponses = transactions/maxItemsPerRequest + 1
         test <@ List.replicate expectedResponses EqxAct.ResponseBackward @ [EqxAct.QueryBackward] = capture.ExternalCalls @>
-        verifyRequestChargesMax 8 // 7.74 // 10.01
+        verifyRequestChargesMax 9 // 8.58 // 10.01
     }
 
-    [<AutoData(MaxTest = 2, SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
+    [<AutoData(MaxFail=1, MaxTest=2, SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
     let ``Can roundtrip against Cosmos, managing sync conflicts by retrying`` ctx initialState = Async.RunSynchronously <| async {
         let log1, capture1 = log, capture
         capture1.Clear()
-        let! conn = connectToSpecifiedCosmosOrSimulator log1
-        // Ensure batching is included at some point in the proceedings
         let batchSize = 3
+        let store = createPrimaryContext log1 batchSize
+        // Ensure batching is included at some point in the proceedings
 
         let context, (sku11, sku12, sku21, sku22) = ctx
         let cartId = % Guid.NewGuid()
 
         // establish base stream state
-        let service1 = Cart.createServiceWithEmptyUnfolds conn batchSize log1
+        let service1 = Cart.createServiceWithEmptyUnfolds store log1
         let! maybeInitialSku =
             let (streamEmpty, skuId) = initialState
             async {
@@ -151,7 +146,7 @@ type Tests(testOutputHelper) =
             do! s4 }
         let log2, capture2 = TestsWithLogCapture.CreateLoggerWithCapture testOutputHelper
         use _flush = log2
-        let service2 = Cart.createServiceWithEmptyUnfolds conn batchSize log2
+        let service2 = Cart.createServiceWithEmptyUnfolds store log2
         let t2 = async {
             // Signal we have state, wait for other to do same, engineer conflict
             let prepare = async {
@@ -192,10 +187,10 @@ type Tests(testOutputHelper) =
     let singleBatchBackwards = [EqxAct.ResponseBackward; EqxAct.QueryBackward]
     let batchBackwardsAndAppend = singleBatchBackwards @ [EqxAct.Append]
 
-    [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
+    [<AutoData(MaxFail=1, SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
     let ``Can correctly read and update against Cosmos with LatestKnownEvent Access Strategy`` value = Async.RunSynchronously <| async {
-        let! conn = connectToSpecifiedCosmosOrSimulator log
-        let service = ContactPreferences.createService log (createCosmosContext conn)
+        let store = createPrimaryContext log 1
+        let service = ContactPreferences.createService log store
 
         let id = ContactPreferences.Id (let g = System.Guid.NewGuid() in g.ToString "N")
         //let (Domain.ContactPreferences.Id email) = id ()
@@ -215,10 +210,10 @@ type Tests(testOutputHelper) =
         test <@ [EqxAct.Tip; EqxAct.Append; EqxAct.Tip] = capture.ExternalCalls @>
     }
 
-     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
+    [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
     let ``Can correctly read and update Contacts against Cosmos with RollingUnfolds Access Strategy`` value = Async.RunSynchronously <| async {
-        let! conn = connectToSpecifiedCosmosOrSimulator log
-        let service = ContactPreferences.createServiceWithLatestKnownEvent (createCosmosContext conn) log CachingStrategy.NoCaching
+        let store = createPrimaryContext log 1
+        let service = ContactPreferences.createServiceWithLatestKnownEvent store log CachingStrategy.NoCaching
 
         let id = ContactPreferences.Id (let g = System.Guid.NewGuid() in g.ToString "N")
         // Feed some junk into the stream
@@ -241,13 +236,13 @@ type Tests(testOutputHelper) =
     let ``Can roundtrip Cart against Cosmos with RollingUnfolds, detecting conflicts based on _etag`` ctx initialState = Async.RunSynchronously <| async {
         let log1, capture1 = log, capture
         capture1.Clear()
-        let! conn = connectToSpecifiedCosmosOrSimulator log1
+        let store = createPrimaryContext log1 1
 
         let context, (sku11, sku12, sku21, sku22) = ctx
         let cartId = % Guid.NewGuid()
 
         // establish base stream state
-        let service1 = Cart.createServiceWithRollingState conn log1
+        let service1 = Cart.createServiceWithRollingState store log1
         let! maybeInitialSku =
             let (streamEmpty, skuId) = initialState
             async {
@@ -280,7 +275,7 @@ type Tests(testOutputHelper) =
             do! s4 }
         let log2, capture2 = TestsWithLogCapture.CreateLoggerWithCapture testOutputHelper
         use _flush = log2
-        let service2 = Cart.createServiceWithRollingState conn log2
+        let service2 = Cart.createServiceWithRollingState store log2
         let t2 = async {
             // Signal we have state, wait for other to do same, engineer conflict
             let prepare = async {
@@ -312,11 +307,11 @@ type Tests(testOutputHelper) =
                 && [EqxAct.Resync] = c2 @>
     }
 
-     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
+    [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
     let ``Can roundtrip against Cosmos, using Snapshotting to avoid queries`` context skuId = Async.RunSynchronously <| async {
-        let! conn = connectToSpecifiedCosmosOrSimulator log
         let batchSize = 10
-        let createServiceIndexed () = Cart.createServiceWithSnapshotStrategy conn batchSize log
+        let store = createPrimaryContext log batchSize
+        let createServiceIndexed () = Cart.createServiceWithSnapshotStrategy store log
         let service1, service2 = createServiceIndexed (), createServiceIndexed ()
         capture.Clear()
 
@@ -341,10 +336,10 @@ type Tests(testOutputHelper) =
 
     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
     let ``Can roundtrip against Cosmos, correctly using Snapshotting and Cache to avoid redundant reads`` context skuId = Async.RunSynchronously <| async {
-        let! conn = connectToSpecifiedCosmosOrSimulator log
         let batchSize = 10
+        let store = createPrimaryContext log batchSize
         let cache = Equinox.Cache("cart", sizeMb = 50)
-        let createServiceCached () = Cart.createServiceWithSnapshotStrategyAndCaching conn batchSize log cache
+        let createServiceCached () = Cart.createServiceWithSnapshotStrategyAndCaching store log cache
         let service1, service2 = createServiceCached (), createServiceCached ()
         capture.Clear()